+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [0.3.2]
 * `PosColumn` can contain encoded text (`textEncoded` field)
 * Bug fix: Columns with `PosTextSize` > `size1`
+* Added Barcode Code128
 
 
 ## [0.3.1]